Refactored TicketForm component
